refactor CC1101 lib clean up code improve speed, reliability and debugging options for rf various bugfixes in command bytes Add IthoCommands: Itho31D9 = 19, Itho31DA = 20, IthoDeviceInfo = 21,

+++ software/NRG_itho_wifi/main/generic_functions.h
@@ -41,5 +41,5 @@ void logLastCommand(const char *command, const char *source);
 void getLastCMDinfoJSON(JsonObject root);
 void updateItho();
 void add2queue();
-void setRFdebugLevel(uint8_t level);
+void setRFdebugLevel(uint16_t level);
 double round(float value, int precision);
